air_bnb/requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Machine Learning Models
catboost>=1.1.0
xgboost>=1.6.0

# Model Persistence
joblib>=1.2.0

# Web Application Framework
streamlit>=1.25.0
streamlit-folium>=0.13.0

# Visualization Libraries
matplotlib>=3.5.0
folium>=0.14.0
shap>=0.41.0

# Environment Variables
python-dotenv>=0.19.0

# OpenAI API
openai>=1.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Additional Utilities
requests>=2.28.0
